feat(core): unify state management for AirPods handling
Refactor to consolidate BLE and AAP data into a unified `PodState` format. Enhance maintainability by centralizing callback mechanisms and improving type consistency across modules.

+++ internal/util/numbers.go
@@ -1,6 +1,6 @@
 package util
 
-func MinOr(a, b *uint8, defaultValue uint8) uint8 {
+func MinOr(a, b *int, defaultValue int) int {
 	if a == nil && b == nil {
 		return defaultValue
 	}
